refactor: migrate GUITAR_TUNINGS import paths to new constants module and remove deprecated guitarTuning type
