+++ frontend/src/__tests__/shared/test-setup.ts
@@ -4,7 +4,7 @@
  */
 import { vi } from 'vitest';
 import { ChordSheet } from '@/types/chordSheet';
-import { GUITAR_TUNINGS } from '@/types/guitarTuning';
+import { GUITAR_TUNINGS } from '@/constants/guitar-tunings';
 import { generateCacheKey as generateChordSheetCacheKey } from '@/cache/core/cache-key-generator';
 
 // Mock localStorage with proper implementation
