+++ frontend/src/hooks/useChordSheet/coordinators/__tests__/fetch-chord-sheet-from-backend.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { fetchChordSheetFromBackend } from '../fetch-chord-sheet-from-backend';
-import { GUITAR_TUNINGS } from '@/types/guitarTuning';
+import { GUITAR_TUNINGS } from '@/constants/guitar-tunings';
 import { getApiBaseUrl } from '@/utils/api-base-url';
 
 // Mock the convert function
