--- frontend/src/types/guitarTuning.ts
@@ -1,2 +1,0 @@
-// Re-export GuitarTuning types from @chordium/types
-export * from '@chordium/types';
